Remove logger handle from CLI Action.run
Remove the logger handle that's passed to the `run` method of all CLI
`Action`s. This should be kept out of the Action interface. It was
previously used to modify the logging level if the `--quiet` option was
specified for some subcommands. Instead, this commit moves the `--quiet`
option to the top-level `splinter` command so that it applies to all
subcommands. This new option is also specified as conflicting with the
verbosity argument.

Signed-off-by: Logan Seeley <seeley@bitwise.io>