Add SQLite support for biome
Adds support for SQLite databases to the biome stores. Operations that
don't require a database that supports the default keyword need no
modifications. Operations that do require a database that supports the
default keyword are re-implemented separately for Postgres and SQLite.

Implementing default keyword operations separately is required because
the `execute` method is provided by different traits for these two
databases, and the compiler needs to know which type of database is
being used so it can decide which trait to use; furthermore, the generic
implementation (using the `SupportsDefaultKeyword` trait bound) cannot
be used together with a `SqliteConnection` implementation, because the
compiler does not know that these implementations cannot overlap.

A side effect of having separate implementations for some operations is
that the store traits need separate implementations as well; again,
this is so the compiler can determine which implementation to use.

This also adds the necessary migrations for SQLite, along with a
migration method (just like PostgreSQL's). There are two differences
between the two sets of migrations: the SQLite migrations do not need
the initial migration (which only has postgres-specific code), and the
`2020-01-10-204453_add_foreign_key_constraint` migration is merged with
the `2019-12-12-200956_create-keys` migration for SQLite (which does not
support adding constraints after a table has been created).

Signed-off-by: Logan Seeley <seeley@bitwise.io>